+++ app/static/index.js
