+++ app/static/login.js
