+++ app/static/register.js
