+++ app/static/utils.js
